refactor: extract mcp types into module
