+++ src/integrations/index.ts
@@ -1,5 +1,10 @@
+export { MCPModule } from "./mcp/mcp.module";
 export { McpToolSourceService } from "./mcp/mcp-tool-source.service";
 export type {
   McpToolSourceDiscovery,
   McpResourceDescription,
-} from "./mcp/mcp-tool-source.service";
+  McpInitializeResult,
+  McpToolDescription,
+  McpToolsListResult,
+  McpResourcesListResult,
+} from "./mcp/types";
