+++ src/integrations/mcp/mcp.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { McpToolSourceService } from "./mcp-tool-source.service";
+
+@Module({
+  providers: [McpToolSourceService],
+  exports: [McpToolSourceService],
+})
+export class MCPModule {}
